requirements.txt
click==8.1.7
pyspellchecker==0.7.2
colorama==0.4.6 